:recycle: Move library to its own directory
